Create new wp-story-editor package and export required modules from story-editor
