+++ package.json
@@ -54,12 +54,14 @@
     "@web-stories-wp/stories-block": "*",
     "@web-stories-wp/stories-carousel": "*",
     "@web-stories-wp/stories-lightbox": "*",
+    "@web-stories-wp/story-editor": "*",
     "@web-stories-wp/templates": "*",
     "@web-stories-wp/text-sets": "*",
     "@web-stories-wp/tinymce-button": "*",
     "@web-stories-wp/tracking": "*",
     "@web-stories-wp/units": "*",
     "@web-stories-wp/widget": "*",
+    "@web-stories-wp/wp-story-editor": "*",
     "@wordpress/api-fetch": "^5.2.1",
     "classnames": "^2.3.1",
     "colorthief": "^2.3.2",
