+++ packages/story-editor/README.md
@@ -1,0 +1,3 @@
+# Story Editor
+
+Web stories core editor
