+++ packages/story-editor/src/index.js
@@ -14,64 +14,19 @@
  * limitations under the License.
  */
 
-/**
- * Internal dependencies
- */
-// The __webpack_public_path__ assignment will be done after the imports.
-// That's why the public path assignment is in its own dedicated module and imported here at the very top.
-// See https://webpack.js.org/guides/public-path/#on-the-fly
-import './publicPath';
+export * from './components/transform';
+export * from './app/config';
+export * from './components/previewPage';
 
-/**
- * External dependencies
- */
-import Modal from 'react-modal';
-import { StrictMode } from 'react';
-import { render } from 'react-dom';
-import { FlagsProvider } from 'flagged';
-import { updateSettings } from '@web-stories-wp/date';
-import { initializeTracking } from '@web-stories-wp/tracking';
+export { default as base64Encode } from './utils/base64Encode';
+export { default as getStoryPropsToSave } from './app/story/utils/getStoryPropsToSave';
+export { default as FontContext } from './app/font/context';
+export { default as useLoadFontFiles } from './app/font/actions/useLoadFontFiles';
+export { default as localStore } from './utils/localStore';
+export { default as StoryPropTypes } from './types';
 
 /**
  * Internal dependencies
  */
 import App from './editorApp';
-import './style.css'; // This way the general editor styles are loaded before all the component styles.
-
-/**
- * Initializes the web stories editor.
- *
- * @param {string} id       ID of the root element to render the screen in.
- * @param {Object} config   Story editor settings.
- * @param {Object} flags    The flags for the application.
- */
-const initialize = (id, config, flags) => {
-  const appElement = document.getElementById(id);
-
-  // see http://reactcommunity.org/react-modal/accessibility/
-  Modal.setAppElement(appElement);
-
-  updateSettings(config.locale);
-
-  initializeTracking('Editor');
-
-  render(
-    <FlagsProvider features={flags}>
-      <StrictMode>
-        <App config={config} />
-      </StrictMode>
-    </FlagsProvider>,
-    appElement
-  );
-};
-
-const initializeWithConfig = () => {
-  const { id, config, flags } = window.webStoriesEditorSettings;
-  initialize(id, config, flags);
-};
-
-if ('loading' === document.readyState) {
-  document.addEventListener('DOMContentLoaded', initializeWithConfig);
-} else {
-  initializeWithConfig();
-}
+export default App;
