+++ packages/wp-story-editor/README.md
@@ -1,0 +1,3 @@
+# WP Story Editor
+
+Web stories editor for WordPress
