+++ packages/wp-story-editor/package.json
@@ -1,0 +1,23 @@
+{
+	"name": "@web-stories-wp/wp-story-editor",
+	"private": true,
+	"engines": {
+		"node": ">= 15",
+		"npm": "~7"
+	},
+	"exports": {
+		"default": "./src/index.js"
+	},
+	"main": "./src/index.js",
+	"type": "module",
+	"dependencies": {
+		"react-modal": "^3.14.3",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
+		"flagged": "^2.0.1",
+		"@web-stories-wp/date": "*",
+		"@web-stories-wp/tracking": "*",
+		"@web-stories-wp/story-editor": "*"
+	},
+	"devDependencies": {}
+}
