+++ packages/wp-story-editor/src/style.css
@@ -1,0 +1,130 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * WordPress admin page integration
+ *
+ * Hides all controls that interfere with the editor UI.
+ */
+
+#screen-meta,
+#screen-meta-links,
+#screen-meta-links ~ :not(.web-stories-wp):not(.clear) {
+  display: none;
+}
+
+body.js.edit-story #wpcontent,
+body.js.edit-story #wpbody-content {
+  padding: 0;
+}
+
+body.edit-story .web-stories-wp {
+  position: relative;
+}
+
+body.js.edit-story #web-stories-editor {
+  background-color: #131516; /* theme.colors.bg.primary */
+  position: relative;
+  height: calc(100vh - 32px); /* ADMIN_TOOLBAR_HEIGHT = 32 */
+  overflow: hidden;
+}
+
+body.js.edit-story #web-stories-editor .loading-message {
+  position: absolute;
+  left: 0;
+  top: 0;
+  right: 0;
+  height: 100%;
+  width: 100%;
+  text-align: center;
+  color: #fff;
+  margin: 0;
+  padding-top: 50px;
+  font-size: 36px;
+  font-family: 'Google Sans', sans-serif;
+}
+
+@keyframes spin-progress {
+  0%,
+  100% {
+    box-shadow: 0em -1.14em 0em 0em #b488fc, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #cbacff, 0.8em 0.8em 0 0em #cbacff,
+      0em 1.14em 0 0em #cbacff, -0.8em 0.8em 0 0em #cbacff,
+      -1.14em 0em 0 0em #cbacff, -0.8em -0.8em 0 0em #cbacff;
+  }
+  12.5% {
+    box-shadow: 0em -1.14em 0em 0em #cbacff, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #b488fc, 0.8em 0.8em 0 0em #cbacff,
+      0em 1.14em 0 0em #cbacff, -0.8em 0.8em 0 0em #cbacff,
+      -1.14em 0em 0 0em #cbacff, -0.8em -0.8em 0 0em #cbacff;
+  }
+  37.5% {
+    box-shadow: 0em -1.14em 0em 0em #cbacff, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #cbacff, 0.8em 0.8em 0 0em #b488fc,
+      0em 1.14em 0 0em #cbacff, -0.8em 0.8em 0 0em #cbacff,
+      -1.14em 0em 0 0em #cbacff, -0.8em -0.8em 0 0em #cbacff;
+  }
+  50% {
+    box-shadow: 0em -1.14em 0em 0em #cbacff, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #cbacff, 0.8em 0.8em 0 0em #cbacff,
+      0em 1.14em 0 0em #b488fc, -0.8em 0.8em 0 0em #cbacff,
+      -1.14em 0em 0 0em #cbacff, -0.8em -0.8em 0 0em #cbacff;
+  }
+  62.5% {
+    box-shadow: 0em -1.14em 0em 0em #cbacff, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #cbacff, 0.8em 0.8em 0 0em #cbacff,
+      0em 1.14em 0 0em #cbacff, -0.8em 0.8em 0 0em #b488fc,
+      -1.14em 0em 0 0em #cbacff, -0.8em -0.8em 0 0em #cbacff;
+  }
+  75% {
+    box-shadow: 0em -1.14em 0em 0em #cbacff, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #cbacff, 0.8em 0.8em 0 0em #cbacff,
+      0em 1.14em 0 0em #cbacff, -0.8em 0.8em 0 0em #cbacff,
+      -1.14em 0em 0 0em #b488fc, -0.8em -0.8em 0 0em #cbacff;
+  }
+  87.5% {
+    box-shadow: 0em -1.14em 0em 0em #cbacff, 0.8em -0.8em 0 0em #cbacff,
+      1.14em 0em 0 0em #cbacff, 0.8em 0.8em 0 0em #cbacff,
+      0em 1.14em 0 0em #cbacff, -0.8em 0.8em 0 0em #cbacff,
+      -1.14em 0em 0 0em #cbacff, -0.8em -0.8em 0 0em #b488fc;
+  }
+}
+body.js.edit-story #web-stories-editor .loading-message:after {
+  content: '';
+  display: block;
+  position: absolute;
+  left: 0;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  margin: 140px auto auto;
+  width: 0.4em;
+  height: 0.4em;
+  border-radius: 50%;
+
+  animation: spin-progress 0.85s infinite
+    cubic-bezier(0.455, 0.03, 0.515, 0.955);
+}
+
+body.edit-story #wpfooter {
+  display: none;
+}
+
+/* Copied from forms.css in WordPress, avoids issues with "Collapse menu" button */
+
+button {
+  font-size: inherit;
+}
